+++ Language/KURE/Combinators/Transform.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE InstanceSigs #-}
 {-# LANGUAGE RankNTypes #-}
 -- |
@@ -54,6 +55,10 @@ import Prelude hiding (id, map, foldr, mapM)
 import Control.Category ((>>>),id)
 import Control.Monad(ap,liftM)
 import Control.Monad.Catch
+#if __GLASGOW_HASKELL__ >= 800
+import qualified Control.Monad.Fail as Fail (fail)
+import Control.Monad.Fail (MonadFail)
+#endif
 
 import Language.KURE.Combinators.Arrow
 import Language.KURE.Combinators.Monad
@@ -232,6 +237,13 @@ instance Monad m => Monad (AnyR m) where
                         return (PBool (b1 || b2) d)
    {-# INLINE (>>=) #-}
 
+#if __GLASGOW_HASKELL__ >= 800
+instance MonadFail m => MonadFail (AnyR m) where
+   fail :: String -> AnyR m a
+   fail = AnyR . Fail.fail
+   {-# INLINE fail #-}
+#endif
+
 instance MonadThrow m => MonadThrow (AnyR m) where
    throwM :: Exception e => e -> AnyR m a
    throwM = AnyR . throwM
@@ -303,6 +315,13 @@ instance Monad m => Monad (OneR m) where
    ma >>= f = OneR $ \ b1 -> do PBool b2 a <- unOneR ma b1
                                 unOneR (f a) b2
    {-# INLINE (>>=) #-}
+
+#if __GLASGOW_HASKELL__ >= 800
+instance MonadFail m => MonadFail (OneR m) where
+   fail :: String -> OneR m a
+   fail msg = OneR (\ _ -> Fail.fail msg)
+   {-# INLINE fail #-}
+#endif
 
 instance MonadThrow m => MonadThrow (OneR m) where
    throwM :: Exception e => e -> OneR m a