+++ Language/KURE/Transform.hs
@@ -44,6 +44,8 @@ import Control.Category
 import Control.Arrow
 
 #if __GLASGOW_HASKELL__ >= 800
+import qualified Control.Monad.Fail as Fail (fail)
+import Control.Monad.Fail (MonadFail)
 import Data.Semigroup (Semigroup(..))
 #endif
 
@@ -132,6 +134,13 @@ instance Monad m => Monad (Transform c m a) where
    fail :: String -> Transform c m a b
    fail = constT . fail
    {-# INLINE fail #-}
+
+#if __GLASGOW_HASKELL__ >= 800
+instance MonadFail m => MonadFail (Transform c m a) where
+   fail :: String -> Transform c m a b
+   fail = constT . Fail.fail
+   {-# INLINE fail #-}
+#endif
 
 -- | Lifting through a Reader transformer, where (c,a) is the read-only environment.
 instance MonadPlus m => MonadPlus (Transform c m a) where