+++ Language/KURE/Walker.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE InstanceSigs #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
@@ -86,6 +87,10 @@ import Control.Arrow
 import Control.Category hiding ((.))
 import Control.Monad
 import Control.Monad.Catch
+#if __GLASGOW_HASKELL__ >= 800
+import qualified Control.Monad.Fail as Fail (fail)
+import Control.Monad.Fail (MonadFail)
+#endif
 
 import Language.KURE.Combinators
 import Language.KURE.Exceptions
@@ -432,6 +437,13 @@ instance (Monoid w, Monad m) => Monad (AllT w m) where
                         return (P d (w1 <> w2))
    {-# INLINE (>>=) #-}
 
+#if __GLASGOW_HASKELL__ >= 800
+instance (Monoid w, MonadFail m) => MonadFail (AllT w m) where
+   fail :: String -> AllT w m a
+   fail = AllT . Fail.fail
+   {-# INLINE fail #-}
+#endif
+
 instance (Monoid w, MonadThrow m) => MonadThrow (AllT w m) where
    throwM :: Exception e => e -> AllT w m a
    throwM = AllT . throwM
@@ -501,6 +513,13 @@ instance Monad m => Monad (OneT w m) where
    ma >>= f = OneT $ do \ mw1 -> do P a mw2 <- unOneT ma mw1
                                     unOneT (f a) mw2
    {-# INLINE (>>=) #-}
+
+#if __GLASGOW_HASKELL__ >= 800
+instance MonadFail m => MonadFail (OneT w m) where
+   fail :: String -> OneT w m a
+   fail msg = OneT (\ _ -> Fail.fail msg)
+   {-# INLINE fail #-}
+#endif
 
 instance MonadThrow m => MonadThrow (OneT w m) where
    throwM :: Exception e => e -> OneT w m a
@@ -575,6 +594,13 @@ instance Monad (GetChild c u) where
                                        (\ e -> GetChild (throwM e) mcu)
                                        kma
    {-# INLINE (>>=) #-}
+
+#if __GLASGOW_HASKELL__ >= 800
+instance MonadFail (GetChild c u) where
+   fail :: String -> GetChild c u a
+   fail msg = GetChild (Fail.fail msg) Nothing
+   {-# INLINE fail #-}
+#endif
 
 instance MonadThrow (GetChild c u) where
    throwM :: Exception e => e -> GetChild c u a