+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * Added `NodeMismatch` (for congruence combinator failures)
  * Added `ConditionalFailure` (a `String` plus `SomeException`)
 * Added `Semigroup` instance for `Transform` (GHC 8.0+)
+* Added `MonadFail` instances (GHC 8.0+)
 
 ## 2.16.12
 * Derive `Typeable` instances